requirements.txt
runpod==1.2.0
torch
diffusers
transformers
accelerate
safetensors
controlnet_aux==0.0.7
xformers
opencv-python
pillow